template and const in ltc moved
